feature/WID-9: Move example page template to core
